Rest post

+++ backend/src/main/java/com/daniel/Java/Spring/Boot/entities/Product.java
@@ -23,7 +23,7 @@ public class Product implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private Double price;
 	
@@ -40,7 +40,7 @@ public class Product implements Serializable {
 		
 	}
 	
-	public Product(Integer id, String name, Double price) {
+	public Product(Long id, String name, Double price) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
@@ -55,11 +55,11 @@ public class Product implements Serializable {
 			return list;
 		}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
